Add discovery of sample bundle torrent from repo
Uses the sample_info repo to pull the latest (or specified) version of the sample bundle.
All the git repo operations are handled within go and the metadata and files are stored in memory. This means the repo should never store any large files as the whole of the repo is stored in memory.

Does not do anything with the torrent file yet.